test digits
Some test digits, which the classifier successfully classified. Also these digits give an indication of what kind of digits it expects. That is, black background, white digit, and soft digit boarders (yes this caused some issues when testing. I assume its because the torchvision MNIST dataset it was trained on all have digits with soft edges)